requirements.txt
Flask==0.10.1
Flask-Sockets==0.1
Jinja2==2.7.1
MarkupSafe==0.18
PyYAML==3.10
Pygments==1.6
Tempita==0.5.1
Unidecode==0.04.14
Werkzeug==0.9.4
argparse==1.2.1
bpython==0.12
decorator==3.4.0
gevent==0.13.8
gevent-websocket==0.3.6
greenlet==0.4.1
gunicorn==18.0
itsdangerous==0.23
oauthlib==0.6.0
pbr==0.5.23
python-dateutil==2.2
python-slugify==0.0.6
requests==2.0.0
requests-oauthlib==0.3.2
six==1.4.1
twython==3.1.0
wsgiref==0.1.2
